Move configurations out of helpers
