+++ web/lib/configuration/arduinosConfig.js
