+++ web/lib/configuration/clientSticksConfig.js
