+++ web/lib/configuration/fakeSensorsConfig.js
